Set up flash-kernel for all Pine64 devices
This commit moves the flash-kernel work into its own recipe which may be
used by other Pine64 recipes. I expect this will need to become more
generic in the future, but not now.

I've removed the work that we did manually before, such as copying DTBs
and compiling boot scripts. That is entirely on flash-kernel now.

I've also created the /etc/flash-kernel/machine file for each device.
This will install the appropriate boot script for the situation.

The boot scripts are installed as part of the device-specific overlay,
while the database is installed as part of the Pine64 overlay.